Rename itemTextToDisplay.ts file
