+++ packages/smart-forms-renderer/src/utils/itemTextToDisplay.ts
